+++ src/components/Header/styles.css
@@ -1,3 +1,3 @@
 header {
-  padding: 30px;
+  padding: 60px 30px;
 }
