+++ src/components/Layout/index.tsx
@@ -2,17 +2,22 @@ import React, { ReactNode } from "react";
 
 import "./styles.css";
 
+import Navbar from "../Navbar";
+
 interface Props {
   children: ReactNode;
 }
 
 const Layout: React.FC<Props> = ({ children }) => {  
   return (
-    <section>
+    <>
+    {/* // <section> */}
+    <Navbar />
       <div id="container">
         { children }
       </div>
-    </section>
+    {/* // </section> */}
+    </>
   );
 };
