+++ src/components/List/styles.css
@@ -1,27 +1,27 @@
-.movies-grid ul{
-  
+#list-container .movies-grid {
+  padding: 0px;
 
-  text-align: center;
+  margin-top: 30px;
+
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 20px;
 }
 
-.movies-grid li {
+#list-container .movies-grid li {
   list-style: none;
-  margin-top: 1rem;
-
-  display: inline-block;
 }
 
-.movies-grid li img {
+#list-container .movies-grid li img {
   cursor: pointer;
 
-  width: 200px;
-  height: 300px;
+  width: 100%;
 
   opacity: 0.7;
   transition: 0.5s;
 }
 
-.movies-grid li img:hover {
+#list-container .movies-grid li img:hover {
   opacity: 1;
 }
 
@@ -45,4 +45,23 @@
 .load-more:hover {
   opacity: 1;
   background-color: aquamarine;
+}
+
+/* small tablet version */
+@media screen and (min-width: 620px) {
+  #list-container .movies-grid {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+/* tablet version */
+@media screen and (min-width: 960px) {
+  #list-container .movies-grid {
+    grid-template-columns: repeat(4, 1fr);
+  }
+}
+/* desktop version */
+@media screen and (min-width: 1200px) {
+  #list-container .movies-grid {
+    grid-template-columns: repeat(5, 1fr);
+  }
 }
