+++ src/components/Main/styles.css
@@ -1,10 +1,14 @@
 main {
     display: flex;
     flex-direction: column;
+    justify-content: center;
+    align-items: center;
 
-    border-top-left-radius: 15px;
-    border-top-right-radius: 15px;
-    background-color: #fcfaf8;
+    width: 100%;
+
+    border-top-left-radius: 30px;
+    border-top-right-radius: 30px;
+    background-color: #222831;
 
     padding: 30px;
 }
