+++ src/pages/Upcoming/index.tsx
@@ -54,7 +54,7 @@ const Upcoming = () => {
 
   return (
     <Layout>
-      <Header>
+      <Header title="Welcome to Awesome Movies">
         <Input name="search" placeholder="Search for any movie" />
       </Header>
