+++ src/styles/global.css
@@ -1,25 +1,36 @@
+/* variabels */
+:root {
+  --primary: #222831;
+  --secondary: #393e46;
+  --tertiary: #00adb5;
+  --quaternary: #eeeeee;
+  --quinary: #828282;
+}
+
+/* body default */
 body {
-  min-height: 100vh;
-  
   display: flex;
   justify-content: center;
-  align-items: center;
 
-  margin: 0;
   box-sizing: border-box;
-  
+  background-color: var(--secondary);
+
   font-family: "Roboto", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
 
+/* reset */
 body,
 input,
 button,
 textarea {
+  margin: 0;
+  padding: 0;
   font: 400 18px Roboto, sans-serif;
+  color: var(--quaternary);
 }
 
-input {
-  color: #828282;
+ul {
+  list-style: none;
 }
